# Pan/externalDB/swagger.yml
swagger: '2.0'
info:
  title: Your API
  version: 1.0.0
paths:
  /get_entity:
    get:
      summary: Get an entity by ID
      operationId: getEntityRoute
      x-swagger-router-controller: app  # Module name
      parameters:
        - name: id
          in: query
          description: ID of the entity to be retrieved
          required: true
          type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Entity not found
  /get_entities_by_time:
    get:
      summary: Get entities by ID and time period
      operationId: getEntitiesByTimeRoute
      x-swagger-router-controller: app  # Module name
      parameters:
        - name: id
          in: query
          description: ID of the entity to be retrieved
          required: true
          type: string
        - name: Year from
          in: query
          description: Start year of entity searching
          required: true
          type: integer
        - name: Month from
          in: query
          description: Start month of entity searching
          required: true
          type: integer
        - name: Day from
          in: query
          description: Start day of entity searching
          required: true
          type: integer
        - name: Hour from
          in: query
          description: Start hour of entity searching
          required: true
          type: integer
        - name: Minute from
          in: query
          description: Start minute of entity searching
          required: true
          type: integer
        - name: Second from
          in: query
          description: Start second of entity searching
          required: true
          type: integer
        - name: Year to
          in: query
          description: End year of entity searching
          required: true
          type: integer
        - name: Month to
          in: query
          description: End month of entity searching
          required: true
          type: integer
        - name: Day to
          in: query
          description: End day of entity searching
          required: true
          type: integer
        - name: Hour to
          in: query
          description: End hour of entity searching
          required: true
          type: integer
        - name: Minute to
          in: query
          description: End minute of entity searching
          required: true
          type: integer
        - name: Second to
          in: query
          description: End second of entity searching
          required: true
          type: integer
      responses:
        '200':
          description: Successful response
        '404':
          description: Entity not found
  /post_entity:
    post:
      summary: Post an entity
      operationId: postEntityRoute
      x-swagger-router-controller: app  # Module name
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: entity
          in: body
          required: true
          schema:
            $ref: '#/definitions/Entity'
      responses:
        '200':
          description: Successful response
  /update_entity:
    patch:  # Adding PATCH method
      summary: Update an entity
      operationId: patchEntityRoute
      x-swagger-router-controller: app  # Module name
      parameters:
        - name: entityId
          in: query
          required: true
          type: string
        - name: atttributesUpdated
          in: body
          description: Object containing the updated attributes and their values
          required: true
          schema:
            $ref: '#/definitions/Entity'
      responses:
        '200':
          description: Successful response  
        '404':
          description: Entity not found
  /delete_entity:
    delete:
      summary: Delete an entity by ID
      operationId: deleteEntityRoute
      x-swagger-router-controller: app  # Module name
      parameters:
        - name: id
          in: query
          description: ID of the entity to be deleted
          required: true
          type: string
      responses:
        '200':
          description: Successful response
        '404':
          description: Entity not found

definitions:
  Entity:
    type: object
    properties:
      id:
        type: string
      type:
        type: string
      version:
        type: string
    required:
      - id
      - type
      - version

  CrowdFlowObserved:
    allOf:
      - $ref: '#/definitions/Entity'
      - type: object
        properties:
          source:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          linkedEntity:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          dateObserved:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          location:
            $ref: '#/definitions/Location'
          peopleCount:
            type: object
            properties:
              type:
                type: string
              value:
                type: number
          congested:
            type: object
            properties:
              type:
                type: string
              value:
                type: boolean

  TrafficViolation:
    allOf:
      - $ref: '#/definitions/Entity'
      - type: object
        properties:
          source:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          linkedEntity:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          dateObserved:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          location:
            $ref: '#/definitions/Location'
          owner:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          description:
            type: object
            properties:
              type:
                type: string
              value:
                type: string

  TransportStation:
    allOf:
      - $ref: '#/definitions/Entity'
      - type: object
        properties:
          name:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          source:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          linkedEntity:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          location:
            $ref: '#/definitions/Location'
          category:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          dateLastReported:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          vehicleLastReported:
            type: object
            properties:
              type:
                type: string
              value:
                type: string

  Vehicle:
    allOf:
      - $ref: '#/definitions/Entity'
      - type: object
        properties:
          category:
            type: string
          source:
            type: string
          linkedEntity:
            type: object
            properties:
              type:
                type: string
              value:
                type: string
          fleetVehicleId:
            type: string
          serviceStatus:
            type: string
          name:
            type: string
          license_plate:
            type: string
          location:
            $ref: '#/definitions/Location'
          fuelFilled:
            type: string

  Location:
    type: object
    properties:
      type:
        type: string
      coordinates:
        type: array
        items:
          type: number